companion-app/requirements.txt
# AI Villager Companion App
# Requirements

# AI
google-generativeai>=0.3.0

# TTS
gTTS>=2.5.0
pygame>=2.5.0

# STT (Speech Recognition)
SpeechRecognition>=3.10.0
PyAudio>=0.2.11

# Networking
websockets>=12.0

# GUI
# tkinter is built-in with Python

# Audio playback
playsound>=1.3.0
